+++ docs/conf.py
@@ -19,7 +19,7 @@ sys.path.insert(0, os.path.abspath('../'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'scikit-hyper'
+project = 'hypers'
 copyright = '2018, Priyank Shah'
 author = 'Priyank Shah'
 
@@ -109,7 +109,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'scikit-hyperdoc'
+htmlhelp_basename = 'hypersdoc'
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -136,7 +136,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'scikit-hyper.tex', 'scikit-hyper Documentation',
+    (master_doc, 'hypers.tex', 'hypers Documentation',
      'Priyank Shah', 'manual'),
 ]
 
@@ -146,7 +146,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'scikit-hyper', 'scikit-hyper Documentation',
+    (master_doc, 'hypers', 'hypers Documentation',
      [author], 1)
 ]
 
@@ -157,8 +157,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'scikit-hyper', 'scikit-hyper Documentation',
-     author, 'scikit-hyper', 'One line description of project.',
+    (master_doc, 'hypers', 'hypers Documentation',
+     author, 'hypers', 'One line description of project.',
      'Miscellaneous'),
 ]