+++ hypers/_view/__init__.py
@@ -1,0 +1,3 @@
+from hypers._view._hsiDialog import hsiPlot
+
+__all__ = ['hsiPlot']